+++ pkgs/test/lib/src/runner/browser/chrome.dart
@@ -194,11 +194,23 @@ Future<WipConnection> _connect(
 }
 
 extension on HttpClient {
+  Encoding determineEncoding(HttpHeaders headers) {
+    final contentType = headers.contentType?.charset;
+
+    /// Using the `charset` property of the `contentType` if available.
+    /// If it's unavailable or if the encoding name is unknown, [latin1] is used by default,
+    /// as per [RFC 2616][].
+    ///
+    /// [RFC 2616]: http://www.w3.org/Protocols/rfc2616/rfc2616-sec3.html
+    return Encoding.getByName(contentType) ?? latin1;
+  }
+
   Future<String?> getString(String url) async {
     final request = await getUrl(Uri.parse(url));
     final response = await request.close();
     if (response.statusCode != HttpStatus.ok) return null;
     var bytes = [await for (var chunk in response) ...chunk];
-    return utf8.decode(bytes);
+    final encoding = determineEncoding(response.headers);
+    return encoding.decode(bytes);
   }
 }