+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.21.1
+
+* Fix a bug loading JS sources with non-utf8 content while parsing coverage
+  information from chrome.
+
 ## 1.21.0
 
 * Allow analyzer version `4.x`.